backend/requirements.txt
fastapi
uvicorn[standard]
pdfminer.six
pytesseract
pillow
httpx
python-multipart
sqlalchemy>=2.0
aiosqlite
